pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bbo-labelgui"
dynamic = ["version"]
description = "GUI for guided data labeling"
readme = "README.md"
license = "BSD"
authors = [
    { name = "BBO-lab @ caesar", email = "kay-michael.voit@caesar.de" },
]
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
dependencies = [
    "bbo-bbo>=0.1.0",
    "bbo_ccvtools",
    "bbo_svidreader",
    "imageio",
    "matplotlib",
    "numpy",
    "tqdm",
]

[project.urls]
Homepage = "https://github.com/bbo-lab/acm-traingui"

[tool.hatch.version]
path = "labelgui/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/labelgui",
]
